Added logout functionality
I imported a Popover icon from material-ui to set the logout prompt as a menuItem for the user to click on to logout. I also created different states of the logout functionality to make the green light in the navbar appear red when clicked on and the user is prompted to logout. Users can now smoothly logout from their session.