Added dynamic params to my profile route
Added dynamic params to my profile route on the front end, so match the id of the currentUser's id. The useParam hook is used to send the userId to the auth/user/find/ on the back end, to fetch the data of the current user to be put into the user profiles information. Now when a user clicks on the profile icon on the navbar, their data gets fetched and used on the profile page, to display their profile info.