Created an Amazon S3 bucket to handle image and video uploads
A Amazon S3 bucket was created to store file and video uploads. The aws.ts file was created on the server side to connect the bucket to the server. Added the proper s3 parameters to the addPost in postControllers, and also included multer for handling the file storage in the middleware. Also refactored the React front end PostForm, to handle image and video uploads. Uploaded media can now be displayed with the appropriate media players.

This commit represents a significant milestone in our application's functionality, allowing users to seamlessly upload, view, and interact with both images and videos. Further improvements and optimizations can be built upon this foundation