refactor: Update notification handling and fix backend query
I created a new table in mysql to hand the notification counts, so they can be incremented or decremented based on if the user clicks on certain icons in the Notifications. I created a count route, controller in the backend, and hooked the count to the notification count in the navbar, and also in the notification.tsx so that if a visibility or trash icon is clicked, it decrements the count. Users can now get rid of their notification count in their navbar.