Created a userSlice.ts in Redux store directory to connect to username and profilePic in Navbar
When a user edited their profile's username or profilePic, the username and profilePic in the Navbar wouldn't update until user logged in. I created a userSlice.ts in the redux store to dispatch login data to the redux, and use that data from the redux to connect to the Navbar. Now, when a user updates their username or profilePic in the Profile.tsx, it automtically updates username and profilePic in Navbar