Created a countSlice in my redux to attempt real time notification count
Notification counts work when posts are liked and commented on, but the state doesn't change in real time, so to solve this I created a countSlice.ts to handle the count state between ViewPost.tsx like and comment mutations, and Navbar.tsx to display the notification count from like and comment mutations happening. I connected both Navbar and ViewPost to countSlice to implement my own real time notification count in the navbar.