Changed the title of the app, and added a new icon to the apps tab
In index.html, I added a new video camera icon to replace the react icon in the tab where the title is. Also changed the title to ClipFlow which is now the name of the app. I also added some corrections to my comments and likes in ViewPost.tsx so that comments and like post requests are sending the right information in the req.body